+++ lib/src/package_analyzer.dart
@@ -214,7 +214,7 @@ Future<String?> _detectGitRoot(String packageDir) async {
     workingDirectory: packageDir,
   );
   if (pr.exitCode == 0) {
-    return pr.stdout.toString();
+    return pr.stdout.toString().trim();
   }
   return null;
 }